proposal: move settings to remote
