+++ client/src/app/index.js
@@ -12,4 +12,3 @@ export { default as App } from './App';
 export { default as AppParent } from './AppParent';
 export { default as KeyboardBindings } from './KeyboardBindings';
 export { default as TabsProvider } from './TabsProvider';
-export { default as Settings } from './Settings';
