+++ client/src/plugins/settings/__tests__/SettingsPluginSpec.js
@@ -14,7 +14,7 @@ import { mount } from 'enzyme';
 
 import SettingsPlugin from '../SettingsPlugin';
 
-import Settings from '../../../app/Settings';
+import Settings from '../../../remote/Settings';
 
 import { Modal } from '../../../shared/ui';
