+++ client/src/remote/Settings.js
