+++ client/src/remote/__tests__/SettingsSpec.js
